// src/shared/models/UserFunction.ts
export default interface UserFunction {
    items: UserFunction[];
    id: number;
    code: string;
    description: string;
    displayName: string;
    parentCode: string;
    appCode: string;
    orders: string;
    host: string;
    prefix: string;
    path: string;
    cssClass: string;
    isEnable: string;
}